move to go 1.11
