+++ cmd/validate/main_test.go
@@ -13,13 +13,13 @@ import (
 	"github.com/golang/protobuf/jsonpb"
 	"time"
 	"fmt"
-	"code.cloudfoundry.org/cf-indicators/pkg/vgo_test"
+	"code.cloudfoundry.org/cf-indicators/pkg/go_test"
 )
 
 func TestValidateIndicators(t *testing.T) {
 	g := NewGomegaWithT(t)
 
-	binPath, err := vgo_test.Build("./")
+	binPath, err := go_test.Build("./")
 	g.Expect(err).ToNot(HaveOccurred())
 
 	t.Run("returns 0 when all metrics are found over 1m", func(t *testing.T) {
