+++ pkg/go_test/build.go
@@ -1,4 +1,4 @@
-package vgo_test
+package go_test
 
 import (
 	"os/exec"
@@ -12,7 +12,7 @@ func Build(packagePath string, args ...string) (compiledPath string, err error) 
 		return "", err
 	}
 
-	build := exec.Command("vgo", "build", "-o", tmpDir + "/build", packagePath)
+	build := exec.Command("go", "build", "-o", tmpDir + "/build", packagePath)
 
 	output, err := build.CombinedOutput()
 	if err != nil {
